Move out ContextMenu logging to self-test code
